+++ BlackoutScanner/Utilities/ServiceLocator.cs
@@ -13,6 +13,8 @@ namespace BlackoutScanner.Utilities
     public static class ServiceLocator
     {
         private static IServiceProvider? _serviceProvider;
+        
+        public static bool IsInitialized => _serviceProvider != null;
 
         public static void Configure()
         {
@@ -45,10 +47,13 @@ namespace BlackoutScanner.Utilities
             services.AddSingleton<IDataManager, DataManager>();
             services.AddSingleton<IGameProfileManager, GameProfileManager>();
             services.AddSingleton<IScreenCapture, ScreenCapture>();
+            services.AddSingleton<IHotKeyManager, HotKeyManager>();
 
             // Register slow-initializing OCR-related services last
             services.AddSingleton<IOCRProcessor, OCRProcessor>();
             services.AddSingleton<IScanner, Scanner>();
+            
+            Log.Information("Service Locator successfully configured all services");
 
             _serviceProvider = services.BuildServiceProvider();
         }